feat: add choosing difficulty and refator from TS to JS
