+++ react-game/package.json
@@ -1,21 +1,14 @@
 {
-  "name": "react-game",
+  "name": "rct-game",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.7.3",
-    "@types/jest": "^26.0.20",
-    "@types/node": "^12.20.4",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
-    "@types/styled-components": "^5.1.7",
+    "@testing-library/user-event": "^12.8.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.3",
-    "styled-components": "^5.2.1",
-    "typescript": "^4.2.2",
     "web-vitals": "^1.1.0"
   },
   "scripts": {
@@ -43,8 +36,7 @@
     ]
   },
   "devDependencies": {
-    "@types/howler": "^2.2.1",
-    "file-loader": "^6.2.0",
+    "styled-components": "^5.2.1",
     "use-sound": "^2.0.1"
   }
 }
