+++ react-game/src/App.module.css
@@ -1,0 +1,27 @@
+html {
+    height: 100%;
+}
+
+body {
+    background-image: url('./img/bcg.jpg');
+    background-size: cover;
+    margin: 0;
+    padding: 0 20px;
+    display: flex;
+    justify-content: center;
+    color: wheat;
+}
+
+* {
+    font-family: 'Catamaran', sans-serif;
+    box-sizing: border-box;
+}
+
+.menu {
+    display: flex;
+    flex-direction: column;
+}
+
+.difficulty {
+    display: none;
+}
