--- react-game/src/App.tsx
@@ -1,131 +1,0 @@
-import React, {useState} from 'react';
-import QuestionCard from '../src/Components/QuestionCard'
-import {Difficulty, fetchQuizQuestions, QuestionState} from "./API";
-import {GlobalStyle} from "./App.styles";
-import useSound from 'use-sound';
-
-const url = "./music/btn.mp3";
-
-const TOTAL_QUESTIONS = 10;
-
-export type AnswerObject = {
-    question: string;
-    answer: string;
-    correct: boolean;
-    correctAnswer: string;
-}
-
-const App: React.FC = () => {
-    const [loading, setLoading] = useState(false);
-    const [questions, setQuestions] = useState<QuestionState[]>([]);
-    const [number, setNumber] = useState(0);
-    const [userAnswers, setUserAnswers] = useState<AnswerObject[]>([]);
-    const [score, setScore] = useState(0);
-    const [gameOver, setGameOver] = useState(true);
-
-    const sound = new Audio(url)
-
-
-
-
-    const startTrivia = async () => {
-        setLoading(true)
-        setGameOver(false)
-
-        const newQuestions = await fetchQuizQuestions(
-            TOTAL_QUESTIONS,
-            Difficulty.EASY
-        );
-
-        setQuestions(newQuestions)
-        setScore(0)
-        setUserAnswers([])
-        setNumber(0)
-        setLoading(false)
-    }
-
-    const checkAnswer = (e: React.MouseEvent<HTMLButtonElement>) => {
-
-        let playPromise = sound.play();
-
-        if (playPromise !== undefined) {
-            playPromise
-                .then(_ => {
-                    console.log("audio played auto");
-                })
-                .catch(error => {
-                    console.log("playback prevented");
-                });
-        }
-
-        if (!gameOver) {
-            const answer = e.currentTarget.value
-
-            const correct = questions[number].correct_answer === answer
-
-            if (correct) setScore(prev => prev + 1)
-
-            const answerObject = {
-                question: questions[number].question,
-                answer,
-                correct,
-                correctAnswer: questions[number].correct_answer,
-            }
-
-            setUserAnswers(prev => [...prev, answerObject])
-        }
-    }
-
-    const nextQuestion = () => {
-        const nextQuestion = number + 1;
-
-        if (nextQuestion === TOTAL_QUESTIONS) {
-            setGameOver(true)
-        } else {
-            setNumber(nextQuestion)
-        }
-    }
-
-
-    return (
-        <>
-            <GlobalStyle/>
-            <h1>REACT QUIZ</h1>
-            {gameOver || userAnswers.length === TOTAL_QUESTIONS ? (
-                <div className="menu">
-                    <button className='start' onClick={startTrivia}>
-                        Start
-                    </button>
-                    <button className=''>
-                        Settings
-                    </button>
-                    <button className=''>
-                        Statistics
-                    </button><button className=''>
-                        Topics
-                    </button>
-                </div>
-
-            ) : null}
-            {!gameOver ? <p className='score'>Score: {score}</p> : null}
-            {loading ? <p>Loading Questions...</p> : null}
-            {!loading && !gameOver && (
-                <QuestionCard
-                    questionNr={number + 1}
-                    totalQuestions={TOTAL_QUESTIONS}
-                    question={questions[number].question}
-                    answers={questions[number].answers}
-                    userAnswer={userAnswers ? userAnswers[number] : undefined}
-                    callback={checkAnswer}
-                />
-            )}
-            {!gameOver && !loading && userAnswers.length === number + 1 && number !== TOTAL_QUESTIONS - 1 ? (
-                <button className='next' onClick={nextQuestion}>
-                    Next Question
-                </button>
-            ) : null}
-        </>
-    );
-};
-
-export default App;
