+++ react-game/src/styles.js
@@ -1,0 +1,8 @@
+export const menu = {
+    display: "flex",
+    flexDirection: "column"
+}
+
+export const difficulty = {
+    display: "none"
+}
