+++ react-game/src/utils.js
@@ -1,0 +1,2 @@
+export const shuffleArray = (array) =>
+    [...array].sort(() => Math.random() - 0.5);
