--- react-game/src/utils.ts
@@ -1,2 +1,0 @@
-export const shuffleArray = (array: any[]) =>
-    [...array].sort(() => Math.random() - 0.5);
